fixed complex128 cast in create_filters(), updated quick_test(), and refactored dsp.py to be more PEP8.

Found an oversight/bug is the dumb in dsp.DSP.create_filters()

Basically, the bp value (now bandpass) was being recast to complex128. GPU are much slower working with complex128 over complex64 values. The problem was solved by preallocating the array bp as np.zeros((n, m), dtype=np.complex64) rather than as an empty list(). I then fill bp by index using enumerate() rather than list().append(). This preserves the complex64 and makes the GPU code run much faster.

Secondly, I switched to CUB and cuTENSOR accelerator backends on Cupy.

Unit test benchmarks show that with this complex128 fix and using the accelerators, develops dsp.py will run 8.77 times faster.
```
# Testing (500 avg)
# System: GPU RTX 2080, CPU i9-9940x 14 Core 3.30 GHz , RAM 32 GB
# |           Name           |    Trail    |  Time Saved  |  Speed Up  |
# | basic with cupy (GPU):   |  133.328 ms |    0.0 ms    |    1.00    |
# | basic with numpy (CPU):  |  776.517 ms | -643.189 ms  |    5.82    |
# | Accelerators cupy (GPU): |  105.813 ms |   27.515 ms  |    1.26    |
# | Fixed dtype cupy (GPU):  |   16.451 ms |  116.877 ms  |    8.10    |
# | Fix + Accel cupy (GPU):  |   15.203 ms |  118.125 ms  |    8.77    |
```