@Lozzy-Bear 2022-05-25
I found the last complex128 call. It happens because we feed a float64 into the GPU. The GPU only wants float16, float32, or complex64 to be fast. From the Cupy documentation:
*FFT functions of NumPy always return numpy.ndarray which type is numpy.complex128 or numpy.float64. CuPy
functions do not follow the behavior, they will return numpy.complex64 or numpy.float32 if the type of the input
is numpy.float16, numpy.float32, or numpy.complex64*

```
# Testing (500 avg)
# System: GPU RTX 2080, CPU i9-9940x 14 Core 3.30 GHz , RAM 32 GB
# |           Name           |    Trial    |  Time Saved  |  Speed Up  |
# | basic with cupy (GPU):   |  133.328 ms |    0.0 ms    |    1.00    |
# | basic with numpy (CPU):  |  776.517 ms | -643.189 ms  |    5.82    |
# | Accelerators cupy (GPU): |  105.813 ms |   27.515 ms  |    1.26    |
# | Fixed dtype cupy (GPU):  |   16.451 ms |  116.877 ms  |    8.10    |
# | Fix + Accel cupy (GPU):  |   15.203 ms |  118.125 ms  |    8.77    |
# | Fix2 + Accel cupy (GPU): |   13.621 ms |  119.707 ms  |    9.79    |
```